// dashboard/src/generated/song_collection_pb.ts
// @generated by protoc-gen-es v2.14.0 with parameter "target=ts"
// @generated from file song_collection.proto (package org.rhythmeta.maimaid.sharing, syntax proto3)
import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file song_collection.proto.
 */
export const file_song_collection: GenFile =
	/*@__PURE__*/
	fileDesc(
		"ChVzb25nX2NvbGxlY3Rpb24ucHJvdG8SHW9yZy5yaHl0aG1ldGEubWFpbWFpZC5zaGFyaW5nImgKE1NvbmdDb2xsZWN0aW9uU2hhcmUSDAoEbmFtZRgBIAEoCRJDCgdlbnRyaWVzGAIgAygLMjIub3JnLnJoeXRobWV0YS5tYWltYWlkLnNoYXJpbmcuU29uZ0NvbGxlY3Rpb25FbnRyeSJOChNTb25nQ29sbGVjdGlvbkVudHJ5Eg8KB3NvbmdfaWQYASABKAkSEgoKY2hhcnRfdHlwZRgCIAEoCRISCgpkaWZmaWN1bHR5GAMgASgJQicKHW9yZy5yaHl0aG1ldGEubWFpbWFpZC5zaGFyaW5nUAG6AgNNTURiBnByb3RvMw",
	);

/**
 * @generated from message org.rhythmeta.maimaid.sharing.SongCollectionShare
 */
export type SongCollectionShare = Message<"org.rhythmeta.maimaid.sharing.SongCollectionShare"> & {
	/**
	 * @generated from field: string name = 1;
	 */
	name: string;

	/**
	 * @generated from field: repeated org.rhythmeta.maimaid.sharing.SongCollectionEntry entries = 2;
	 */
	entries: SongCollectionEntry[];
};

/**
 * Describes the message org.rhythmeta.maimaid.sharing.SongCollectionShare.
 * Use `create(SongCollectionShareSchema)` to create a new message.
 */
export const SongCollectionShareSchema: GenMessage<SongCollectionShare> = /*@__PURE__*/ messageDesc(file_song_collection, 0);

/**
 * @generated from message org.rhythmeta.maimaid.sharing.SongCollectionEntry
 */
export type SongCollectionEntry = Message<"org.rhythmeta.maimaid.sharing.SongCollectionEntry"> & {
	/**
	 * @generated from field: string song_id = 1;
	 */
	songId: string;

	/**
	 * @generated from field: string chart_type = 2;
	 */
	chartType: string;

	/**
	 * @generated from field: string difficulty = 3;
	 */
	difficulty: string;
};

/**
 * Describes the message org.rhythmeta.maimaid.sharing.SongCollectionEntry.
 * Use `create(SongCollectionEntrySchema)` to create a new message.
 */
export const SongCollectionEntrySchema: GenMessage<SongCollectionEntry> = /*@__PURE__*/ messageDesc(file_song_collection, 1);